Re-implement fix to eliminate the redundant redraw
The fix for the redundant redraw when typing in the editor, as first
attempted in 4987c25, succeeded in eliminating the extra redraw.
However, it also caused the summary view to become out of sync with the
editor contents when switching dates. This commit is a second attempt at
eliminating the extra redraw, which it succeeds in doing, and without
any unwelcome side effects.