Use UTC for log times to ignore DST completely
On the date of this writing, the DST shift just occurred. This caused my
"midnight crossover" test to fail by a single-hour difference. To fix
this, we should force all Moment objects (used for log times) to be UTC
(which effectively makes all times DST-agnostic).

The calendar and date components still use local time because an actual
date is involved. We only want to use UTC where a date-agnostic time is
displaying.