Maximize tree-shaking in autocompletion worker
Using _.chain() can effectively disable tree-shaking, since the compiler
is unable to tree-shake methods that aren't part of the same namespace
(the `import * as _` syntax). Also replace _.toPairs() with
Object.entries() to reduce dependence on Lodash in favor of native
functionality. This reduces the end file size of the worker bundle from
~90 KB to ~20 KB, a substantial improvement.