Migrate away from dedicated Preferences test file
Instead, each preference will be managed in the test file that it most
relates to. For instance, the behavior of changing the autocomplete
preference will be tested inside the autocomplete test file. This is
because testing the behavior of changing a preference is far more
important than the trivial toggling of a preference value.