Fix next completion not canceling when deleting trailing space
This fixes a bug where, when completing a word and then pressing the
Delete key to delete the trailing space, the completion placeholder was
still active. Ironically, this was due to the cancel() call in the
EditorComponent.autocomplete() method; the cancel was resetting the
autocompleter model's (last-computed) 'completionQuery' property, which
was causing the substring optimization in fetchCompletions() to run.

The fix seems to be removing the cancel() call, but then (crucially)
resetting the rendered autocomplete in the DOM to prevent a flicker of
the previous placeholder text when tab-completing past the initial
autocomplete run.

With this fix, the behavior of canceling the autocomplete property in
this scenario is not only desired, but matches the behavior of the
previous autocomplete rendering implementation.